feat : Đăng nhập bằng google #62

+++ src/services/Comment/delete.js
@@ -2,13 +2,18 @@
 import { StatusCodes } from 'http-status-codes'
 import ApiError from '../../utils/ApiError.js'
 import Comment from '../../model/Comment.js'
-import { commentRecursiveService } from '../CommentRecursive/index.js'
 import commentRecursive from '../../model/commentRecursive.js'
 export const removeService = async (reqBody) => {
   try {
     const id = reqBody.params.id
     // Tìm đối tượng Food theo ID trước khi cập nhật
-
+    const commentCurrent = await Comment.findById(id)
+    if (commentCurrent.comments.length > 0) {
+      throw new ApiError(
+        StatusCodes.CONFLICT,
+        'This comment has sub children. Delete sub children before delete this comment!'
+      )
+    }
     // Cập nhật trường isDeleted thành true để đánh dấu xóa mềm
     const data = await Comment.findByIdAndDelete(id)
